Fixed some display issues, added some docs, working on having gear properteis scale with rating.

+++ SR5Builder/Loaders/AugmentLoader.cs
@@ -13,7 +13,7 @@ namespace SR5Builder.Loaders
 
         public string Target { get; set; }
 
-        public float[] Bonus { get; set; }
+        public decimal[] Bonus { get; set; }
 
         public Augment ToAugment(IAugmentContainer owner)
         {
